requirements.txt
websockets==11.0.3
folium==0.19.2
streamlit_folium==0.23.2
pymongo==4.10.1
